# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Farm Market API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Farmer {
  id: number;
  name: string;
  /** @nullable */
  bio?: string | null;
  location: string;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export interface FarmerInput {
  /** @minLength 1 */
  name: string;
  bio?: string;
  /** @minLength 1 */
  location: string;
  imageUrl?: string;
}

export interface FarmerUpdate {
  /** @minLength 1 */
  name?: string;
  bio?: string;
  location?: string;
  imageUrl?: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  icon?: string | null;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  icon?: string;
}

export interface CategorySummary {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  icon?: string | null;
  productCount: number;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  unit: string;
  /** @nullable */
  imageUrl?: string | null;
  farmerId: number;
  /** @nullable */
  farmerName?: string | null;
  /** @nullable */
  farmerLocation?: string | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  inStock: boolean;
  featured?: boolean;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  price: number;
  /** @minLength 1 */
  unit: string;
  imageUrl?: string;
  farmerId: number;
  categoryId: number;
  inStock?: boolean;
  featured?: boolean;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  price?: number;
  unit?: string;
  imageUrl?: string;
  categoryId?: number;
  inStock?: boolean;
  featured?: boolean;
}

export interface OrderItem {
  productId: number;
  /** @nullable */
  productName?: string | null;
  quantity: number;
  priceAtOrder: number;
  /** @nullable */
  unit?: string | null;
}

export interface Order {
  id: number;
  buyerName: string;
  buyerEmail: string;
  status: string;
  totalAmount: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  items: OrderItem[];
}

export interface OrderItemInput {
  productId: number;
  quantity: number;
}

export interface OrderInput {
  /** @minLength 1 */
  buyerName: string;
  /** @minLength 1 */
  buyerEmail: string;
  notes?: string;
  items: OrderItemInput[];
}

export interface OrderStatusUpdate {
  status: string;
}

export interface MarketStats {
  totalFarmers: number;
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface LocalRegistrationInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  firstName: string;
  /** @maxLength 80 */
  lastName?: string;
  /** @maxLength 320 */
  email: string;
  /**
     * @minLength 10
     * @maxLength 128
     */
  password: string;
}

export interface LocalLoginInput {
  /** @maxLength 320 */
  email: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  password: string;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type ListProductsParams = {
/**
 * @nullable
 */
categoryId?: number | null;
/**
 * @nullable
 */
farmerId?: number | null;
/**
 * @nullable
 */
search?: string | null;
};

export type ListOrdersParams = {
/**
 * Filter orders by buyer email
 */
buyerEmail?: string;
};