# lib/api-zod/src/generated/types/localLoginInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Farm Market API
 * OpenAPI spec version: 0.1.0
 */

export interface LocalLoginInput {
  /** @maxLength 320 */
  email: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  password: string;
}