# lib/api-zod/src/generated/types/localRegistrationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Farm Market API
 * OpenAPI spec version: 0.1.0
 */

export interface LocalRegistrationInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  firstName: string;
  /** @maxLength 80 */
  lastName?: string;
  /** @maxLength 320 */
  email: string;
  /**
     * @minLength 10
     * @maxLength 128
     */
  password: string;
}